requirements.txt
streamlit
PIL
azure-cognitiveservices-vision-computervision
pillow